[Coroutines][NFC] Refactor CoroCloner
* Move CoroCloner to its own header. For now the header is located in
  lib/Transforms/Coroutines
* Change private to protected to allow inheritance
* Create CoroSwitchCloner and move some of the switch specific code into
  this cloner.
* This continus previous work done to CoroFrame. The goal is to isolate
  the cloner code required for each ABI as much as is reasonable. The
  base class should include a useful set of helpers and base
  implementations. More changes will follow.
